+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `[jest-reporters]` Fixes notify reporter on Linux (using notify-send) ([#10393](https://github.com/facebook/jest/pull/10400))
 - `[jest-snapshot]` Correctly handles arrays and property matchers in snapshots ([#10404](https://github.com/facebook/jest/pull/10404))
+- `[jest-jasmine2, jest-circus]` Find correct location for `test.each` tests ([#10413](https://github.com/facebook/jest/pull/10413))
 
 ### Chore & Maintenance